[feat] 로그인 시 사용자 정보 가져오기 구현

+++ src/main/java/com/moneyplan/expense/service/ExpenseService.java
@@ -34,13 +34,11 @@ public class ExpenseService {
     private final CategoryRepository categoryRepository;
 
     @Transactional
-    public ExpenseRes createExpense(ExpenseReq req) {
+    public ExpenseRes createExpense(Member member, ExpenseReq req) {
 
-        // 회원가입/로그인 구현 전 임시 코드
-        Member member = memberRepository.findById(1L)
-            .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-        // 임시 코드 끝
-
+        if (member == null) {
+            throw new BusinessException(ErrorCode.MISSING_AUTHENTICATION);
+        }
         Category category = categoryRepository.findById(req.getCategoryId())
             .orElseThrow(() -> new BusinessException(ErrorCode.CATEGORY_NOT_FOUND));
 
@@ -51,15 +49,9 @@ public class ExpenseService {
     }
 
     @Transactional
-    public ExpenseRes updateExpense(Long id, ExpenseReq req) {
+    public ExpenseRes updateExpense(Member member, Long id, ExpenseReq req) {
 
-        // 회원가입/로그인 구현 전 임시 코드
-        Member member = memberRepository.findById(1L)
-            .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-        // 임시 코드 끝
-
-        Expense expense = expenseRepository.findById(id)
-            .orElseThrow(() -> new BusinessException(ErrorCode.EXPENSE_NOT_FOUND));
+        Expense expense = verifyAccess(member, id);
 
         Category category = categoryRepository.findById(req.getCategoryId())
             .orElseThrow(() -> new BusinessException(ErrorCode.CATEGORY_NOT_FOUND));
@@ -70,40 +62,24 @@ public class ExpenseService {
     }
 
     @Transactional
-    public void deleteExpense(Long id) {
-
-        // 회원가입/로그인 구현 전 임시 코드
-        Member member = memberRepository.findById(1L)
-            .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-        // 임시 코드 끝
-
-        Expense expense = expenseRepository.findById(id)
-            .orElseThrow(() -> new BusinessException(ErrorCode.EXPENSE_NOT_FOUND));
-
+    public void deleteExpense(Member member, Long id) {
+        verifyAccess(member, id);
         expenseRepository.deleteById(id);
     }
 
     @Transactional
-    public ExpenseRes getExpense(Long id) {
-
-        // 회원가입/로그인 구현 전 임시 코드
-        Member member = memberRepository.findById(1L)
-            .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-        // 임시 코드 끝
-
-        Expense expense = expenseRepository.findById(id)
-            .orElseThrow(() -> new BusinessException(ErrorCode.EXPENSE_NOT_FOUND));
-
+    public ExpenseRes getExpense(Member member, Long id) {
+        Expense expense = verifyAccess(member, id);
         return ExpenseRes.of(expense);
     }
 
     @Transactional
-    public ExpensesRes getExpenses(int page, int size, ExpenseFilter filter) {
+    public ExpensesRes getExpenses(Member member, int page, int size, ExpenseFilter filter) {
 
-        // 회원가입/로그인 구현 전 임시 코드
-        Member member = memberRepository.findById(1L)
-            .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-        // 임시 코드 끝
+        if (member == null) {
+            throw new BusinessException(ErrorCode.MISSING_AUTHENTICATION);
+        }
+        filter.updateMemberId(member.getId());
 
         Pageable paging = PageRequest.of(page, size);
         Page<Expense> expensePage = expenseRepository.findWithFilters(paging, filter);
@@ -122,5 +98,18 @@ public class ExpenseService {
             .totalAmount(totalAmount)
             .expenseCategoryTotals(categoryTotals)
             .build();
+    }
+
+    private Expense verifyAccess(Member member, Long expenseId) {
+        if (member == null) {
+            throw new BusinessException(ErrorCode.MISSING_AUTHENTICATION);
+        }
+        Expense expense = expenseRepository.findById(expenseId)
+            .orElseThrow(() -> new BusinessException(ErrorCode.EXPENSE_NOT_FOUND));
+
+        if (member.getId() != expense.getMember().getId()) {
+            throw new BusinessException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+        return expense;
     }
 }
